refactor(substrate): split mail registry dispatch and addressing concerns
